bus/arm-system-profiler: Basic driver for the ARM System Profiler
This patch implements a basic driver for the ARM System Profiler

The System Profiler registers are exposed symbolically via debugfs,
allowing root do manipulate and experiment with the profiler.
If an interrupt is defined for the profiler in the device tree and
ftrace is enabled, then captured data can be streamed out via trace
events.

The current user interface is minimal, with virtually no abstraction.

Future versions of the driver may substantially change the interface,
so it should be considered experimental for now.

Signed-off-by: Dave Martin <Dave.Martin@arm.com>