Juno: Change I2C bus speed used for HDMI to 100kHz.
The I2C bus driver has problems under heavy load to keep the FIFO
filled. Combined with the auto-STOP setting baked into the hardware,
it spells disaster for conversations with HDMI chip, specially
around the boot time when we want to retrieve the EDID information.

Change the bus speed to 100kHz to increase the chance of servicing
the "FIFO empty" interrupt in time.

Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>