Juno: defconfig: Remove the reduced debug info from defconfig and add SLUB debugging.
The reduced debug info trips the debugging tools that rely on certain
information being present in the debug build in order to figure out
the OS support structures. Also, SLUB debugging helps trap a corner
case where a race condition exists in the SLUB allocator when is
updating the sysfs statistics information and it races with a
SLUB deallocator over maintainance of the freelist pointer.

Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>