drm: rockchip: Don't pass DRM fake offset to dma-api
Set vm_pgoff to 0 after using it to look up the GEM node, before passing
it on rockchip_gem_mmap_buf() where the offset must be from the start of
the buffer.

Passing in the fake offset currently works because the
dma_mmap_attrs implementation that is used for this device,
arm_iommu_mmap_attrs, ignores the offset completely.

Signed-off-by: Ørjan Eide <orjan.eide@arm.com>
Signed-off-by: Guillaume Tucker <guillaume.tucker@arm.com>