PM / OPP: add a function to get the voltage for disabled OPPs
The OPP library is now used for power models, to calculate the power
that a device would consume at a specific OPP.  To do that, we use a
simple power model which takes frequency and voltage as inputs.  We get
the voltage and frequency from the OPP library.

A devfreq cooling device for the thermal framework controls temperature
by disabling OPPs.  The power model needs to calculate the power that
would be consumed if we reenabled the OPP.  dev_pm_opp_get_voltage()
doesn't work for disabled OPPs.

Add a dev_pm_opp_get_voltage_always() that works both for enabled and
disabled OPPs to be used by the power model.  The documentation for this
function clearly states that you should use dev_pm_opp_get_voltage()
instead unless you know what you're doing.

Cc: "Rafael J. Wysocki" <rjw@rjwysocki.net>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Javi Merino <javi.merino@arm.com>

[Updated for OPPv2, i.e. new file location and add opp_rcu_lockdep_assert]
Signed-off-by: Jon Medhurst <tixy@linaro.org>