arm64: psci: add suspend_set_ops functions
Kernel-3.18 is missing suspend_set_ops. This is required
by the power management framework to suspend the system to RAM.
Kernel will call all the devices suspend calls that are registered
to suspend framework. If any one of the device doesn't go into
suspend state then the whole system will not be suspend.

This is a workaround to fix this issue. SYSTEM_SUSPEND should be
handled by PSCI code which will issue smc call with FunctionID 000E

Jira: MPDBSP-402 - Add suspend_set_ops in kernel-3.18
Change-Id: I4e212582c79efbe3e7bc12e37c1490438f3a78d8