drm/rockchip: Send events for same-fb flips
    Previous message: [PATCH 1/2] drm/rockchip: Use CRTC vblank event interface
    Next message: [PATCH 2/2] drm/rockchip: Send events for same-fb flips
    Messages sorted by: [ date ] [ thread ] [ subject ] [ author ]

Rockchip previously treated a pageflip to the same framebuffer as a
no-op, discarding the event if one was requested. This breaks Weston,
which, when idle, sends a no-op vblank event to discover vblank
timings if the vblank query interface is not usable.

Silently dropping events is also quite a hostile thing to do to
userspace in general.

Signed-off-by: Daniel Stone <daniels at collabora.com>
Cc: Sjoerd Simons <sjoerd.simons at collabora.co.uk>
Cc: Heiko Stuebner <heiko at sntech.de>
Signed-off-by: Guillaume Tucker <guillaume.tucker@arm.com>