drivers/base: enable dma_buf_test_exporter support
Enable support for dma_buf_test_exporter. This is needed to run
Wayland with fbdev backend (allocator option in scons). To do this
Kconfig and Makefile need to be modified in order to properly include
dma_buf_test_exporter directory. DMA_SHARED_BUFFER_TEST_EXPORTER
needs to be enabled as well.

Jira: MPDBSP-457
Change-Id: Iff99546bb530bb09cb582475f4a934054abb5e93