arm64: juno: dts: add dtsi for HDMI daughterboard
The V2C-HDMI board is a daughterboard which plugs into the LogicTile on
a Juno development board, providing two HDMI outputs for use by devices
instantiated in the FPGA.

Add a device-tree for this daughterboard. It should be included as a
child of the i2c bus it is connected to.

Change-Id: Ia2b887972c4a2191d77c131ce602b491a8850c72
Signed-off-by: Brian Starkey <brian.starkey@arm.com>