arm64: juno: dts: add dtsi for HDMI daughterboard
The V2C-HDMI board is a daughterboard which plugs into the LogicTile on
a Juno development board, providing two HDMI outputs for use by devices
instantiated in the FPGA.

Add a device-tree for this daughterboard. It should be included as a
child of the i2c bus it is connected to.

Change-Id: I29da99fd731bd38914711024474a4ed15925fc9a
Signed-off-by: Brian Starkey <brian.starkey@arm.com>