arm64: dts: add Juno dts with Mali GPU variants
Default Juno dt file has only Thor (Mali-T624 SoC) GPU node.  We
also need to support different FPGA GPUs that will be installed
on the board with GPU architecture specific device
addresses. This change keeps the Juno SoC GPU option and also
adds FPGA GPU variants on Juno r0 and r1.

The supported GPU configurations are:

  * Mali-T604 from the Juno SoC
  * Mali Midgard on FPGA
  * Mali Bifrost on FPGA

System coherency needs to be set to ACELite for Midgard, and both
ACE and ACELite for Bifrost on Juno r1.

The value of the system-coherency property is defined as follos:

  * 0:  ACELite
  * 1:  ACE
  * 31: disabled (default set in hardware)

For configurations which do not support system coherency at all,
we leave the value to the default as set in hardware and do not
include system-coherency in the device tree.

The resulting list of device trees with system-coherency settings is:

  * juno-r0.dtb                 (ACELite)
  * juno-r0-midgard.dtb         (None)
  * juno-r0-bifrost.dtb         (None)
  * juno-r1.dtb                 (ACELite)
  * juno-r1-midgard.dtb         (None)
  * juno-r1-midgard-acelite.dtb (ACELite)
  * juno-r1-bifrost.dtb         (None)
  * juno-r1-bifrost-acelite.dtb (ACELite)
  * juno-r1-bifrost-ace.dtb     (ACE)

Also, fix clock name using "clk_mali" instead of "vgpu".

Jira: MPDBSP-547
Change-Id: I965f7cfa5db01bdcd683f39bf6771c3541a1e3e6
Signed-off-by: Guillaume Tucker <guillaume.tucker@arm.com>