arm64: dts: add fpga_sysctl for Mali Bifrost GPUs
Add fpga_sysctl node in all dts for Bifrost on Juno r1 and r2.
This only makes a difference on Heimdall and more recent Bifrost
GPUs.  It does not break anything with Mimir so it can be enabled
on all Bifrost variants.  Heimdall can only be used on Juno r1 or
r2, which is why this change is not in the Juno r0 dts.

Jira: MPDBSP-895
Change-Id: I85fdfb934756afdc280f5fe0668254fd7b9df2d1