PM / OPP: add LSK macro to detect API back-port
As the OPPv2 API has been back-ported from mainline onto the LSK
3.18 branch, add a macro to show this so it can be used in
out-of-tree modules to build against this tree.

The main reason to use this is for the drivers/gpu/arm/midgard
GPU kernel driver when built out-of-tree as checking kernel
versions is not enough to determine which OPP API is present in
the kernel tree.

Jira: MPDBSP-960
Change-Id: Ie5a674b54911dbf12bb9dcf383dab259abda6fee