linaro/configs: set selinux as default security module
AppArmor is the default security module in LSK(Linaro Stable Kernel).
This will work on most of linux distribution. But to work on android
with androidboot.selinux=permission it needs to be set to selinux. In
other words DEFAULT_SECURITY config should be set to "selinux".

Setting CONFIG_DEFAULT_SECURITY="selinux" will enabled kernel to create
selinxfs and mount this filesytem. Android will then read
/sys/fs/selinux for selinux related attributes, permissions and
policies.

This change is to remove apparmor as default security module and make
selinux as default security module.

Jira: MPDBSP-1084
Change-Id: Ib9251ceabcffb59e87f3ef366ece5f952e7f6fbb