arm64: dts: disable cpuidle states for Juno R0 & R1
The cpuidle framework does not handle cpuidle states properly,
there appears to be a bug in the CPU framework or in cpuidle
driver. This is leading to performance degradation and as a
result system tests and conformance tests are timing out.

Disabling cpuidle states for all CPUs and clusters for now until
we have a proper fix. This means CPUs do not register with the
cpuidle framework and as a result CPUs do not have cpuidle
states.

While switching kernel versions from 3.18.20 to 3.18.34, for some
reason we omitted this change for Juno R0 and R1, instead we
only disabled cpuidle states for Juno R2, see commit:
  b043e1e94ee989b0818624d46f8304dad0a26b16

This caused some issues in TI2 mali_cl_unit regression testing.
See following TI2 jobs:
  http://mpdti2.euhpc.arm.com/tables/job/27565717
  http://mpdti2.euhpc.arm.com/tables/job/28010593

Jira: MPDBSP-1245
Change-Id: I2f528b4271ce5c22a21aaa07da99e215133c0f75