drm: mali-dp: add inverse-gamma support
Inverse gamma in DP500 is fairly straightforward: there is one IG table
for the video layer and one table for the first graphics layer.
DP550 & DP650 have 3 tables in hardware, and every layer can reference
any of the 3 tables. Table selection can be done atomically, but
re-writing them requires a modeset due to the lack of shadow registers.

Signed-off-by: Mihail Atanassov <mihail.atanassov@arm.com>
[
  Update to use plane properties found here
  https://lists.freedesktop.org/archives/dri-devel/2018-March/168885.html
  - minor cleanups
  - fix igamma disable flow
]
Signed-off-by: Alexandru Gheorghe <alexandru-cosmin.gheorghe@arm.com>

Change-Id: I1edabbd0dc17c208b72d4f030d2209c85012a97e