arm64: dts: add DT for on FPGA Mali GPU on ARMltd Juno platform
Default Juno dt file has only Thor (Mali-T624 SoC) GPU node. We also need to
support different FPGA GPUs that will be installed on the board with GPU
architecture specific device addresses. This change keeps the Juno SoC GPU
option and also adds FPGA GPU variants on Juno r0, r1 and r2.

The supported GPU configurations are:

  * Mali-T624 from the Juno SoC
  * Mali Midgard on FPGA
  * Mali Bifrost on FPGA

System coherency needs to be set to ACELite for Midgard, and both ACE and
ACELite for Bifrost on Juno r1 and r2.

The value of the system-coherency property is defined as follows:

  * 0:  ACELite
  * 1:  ACE
  * 31: disabled (default set in hardware)

For configurations which do not support system coherency at all, we leave the
value to the default as set in hardware and do not include system-coherency in
the device tree.

The resulting list of device trees with system-coherency settings is:

  * juno-r0.dtb                 (ACELite)
  * juno-r0-midgard.dtb         (None)
  * juno-r0-bifrost.dtb         (None)

  * juno-r1.dtb                 (ACELite)
  * juno-r1-midgard.dtb         (None)
  * juno-r1-midgard-acelite.dtb (ACELite)
  * juno-r1-bifrost.dtb         (None)
  * juno-r1-bifrost-ace.dtb     (ACE)
  * juno-r1-bifrost-acelite.dtb (ACELite)

  * juno-r2.dtb                 (ACELite)
  * juno-r2-midgard.dtb         (None)
  * juno-r2-midgard-acelite.dtb (ACELite)
  * juno-r2-bifrost.dtb         (None)
  * juno-r2-bifrost-ace.dtb     (ACE)
  * juno-r2-bifrost-acelite.dtb (ACELite)

This work is a merge of the following patches from branch
linux-4.9-android-common-a:

  * 5dea30bf0dd47df05b877b06d728b94b1375bc96
  * c3025a0c3633a35bc7dec463598cb52b9881bd4d
  * 62dd568f3bc1c1eb9c4c6ec17d6d2c6c82c17f1d

Jira: MPDBSP-2055
Change-Id: I775a9faba183d21701c6366a582797ca14a94773
Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>
(cherry picked from commit ec979dff52cc5a1f8c36b17dfb60ded10ff35f74)
[steven.wang@arm.com: enabling ACELite for Juno R2]
Signed-off-by: Steven Wang <steven.wang@arm.com>