video: adf: add adf_sync to support adf on kernel4.9
The new drivers/dma-buf/sw_sync.c implements most of the same
functionality that the old sw_sync had, but it is not exported.
We can copy the code from this file under drivers/video/adf/ and
implement on top of it. This will provide the same interface to the
 ADF framework code, as well as to the malidp-v0 driver. This is a
similar approach to what has been done by drivers/staging/goldfish/
goldfish_sync_*

Change-Id: Ifd2b31f58b8f3041ffd0de5f4cb265a54986e2f3
Signed-off-by: Dennis Lu <dennis.lu@arm.com>
Signed-off-by: Brian Starkey <brian.starkey@arm.com>