drivers/misc: Add support for Mali-DP reset widget
This adds support for the Mali-DP reset widget included in some Mali-DP
development bitfiles. The driver exposes an interface in sysfs which can
be used to reset the Mali-DP hardware, and integrates with the system PM
framework to conduct a reset during the "suspend_noirq" phase.

The sysfs interface consists of 2 files:
 - mask: The reset bit mask. The bottom 3 bits are meaningful (and the rest
   will be ignored):
      bit 2: Global reset
      bit 1: Primary core reset
      bit 0: Secodary core reset
   N.B. Setting "mask" to zero effectively disables the reset widget
 - reset: Writing any value to this file will conduct a reset according
   to the bits set in "mask"

Change-Id: I028ed7a5e76f77da949d79ff9c48042d2ba12639