drm: mali-dp: Log internal errors v2
Status register contains a lot of bits for reporting internal errors
inside Mali DP. Currently, we just silently ignore all of the erorrs,
that doesn't help when we are investigating different bugs, especially
on the FPGA models which have a lot of constrains, so we could easily
end up in AXI or underrun errors.

Add a new KConfig option called CONFIG_DRM_MALI_DISPLAY_DEBUG, which
enables the logging of the errors present in the status register.

Errors will be reported in  /sys/kernel/debug/tracing/trace,
still not noisy enough, but better than silently ignoring them.

E.g:
<idle>-0     [000] d.h1  2387.240042: malidp_de_irq: error occurred at vblank 596 DE_STATUS is 0x00000001
<idle>-0     [000] d.h1  2387.256703: malidp_de_irq: error occurred at vblank 597 DE_STATUS is 0x00000001
<idle>-0     [000] d.h1  2387.273458: malidp_de_irq: error occurred at vblank 598 DE_STATUS is 0x00000001
<idle>-0     [000] d.h1  2387.290035: malidp_de_irq: error occurred at vblank 599 DE_STATUS is 0x00000001

In addition to that, I removed the setting of MALIDP550_SE_IRQ_AXI_ERR
bit in the irq_mask, since that bit doesn't exist.

v2: Consistently refer the hardware as Mali DP.

Acked-by: Liviu Dudau <liviu.dudau@arm.com>
Signed-off-by: Alexandru Gheorghe <Alexandru-Cosmin.Gheorghe@arm.com>
Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>