arm64: dts: juno: GPIO buttons as active low
The Juno's GPIO buttons are active low and should be set as such. The previous
value of 0x4 is unknown as the current GPIO_LINE_OPEN_DRAIN was not available
when those DTs have been introduced (4.2).

This is not possible to bring into juno-motherboard.dtsi the common include:

  include/dt-bindings/gpio/gpio.h

Jira: MPDBSP-1796
Change-Id: I5cd24abbeb21843228e7885c038b815807bb4f63
Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>
Signed-off-by: Dennis Lu <dennis.lu@arm.com>