drm: Add component-aware simple encoder
This is a simple DRM encoder that gets its connector timings information
from a OF subnode in the device tree and exposes that as a "discovered"
panel. It can be used together with component-based DRM drivers in an
emulated environment where no real encoder or connector hardware exists
and the display output is configured outside the kernel.

The patch is original from:
  * From cc81b70e3fb005abf2f331a6352ca1a4b76dde34 Mon Sep 17 00:00:00 2001
  * From: Liviu Dudau <Liviu.Dudau@arm.com>
  * Date: Mon, 7 Mar 2016 10:55:26 +0000

Change-Id: If1240ac8e352363cfc623a38943ae4aee31d6680
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>