drm: Fix crtc color management when doing suspend/resume
When doing suspend/resume drivers usually use the
drm_atomic_helper_suspend/drm_atomic_helper_resume pair for saving the
state and then re-comitting it.

The problem is that drm_crtc_state has a bool field called
color_mgmt_changed, which mali-dp and other drivers uses it to detect
if coefficients need to be reprogrammed, but that never happens
because the save state has color_mgmt_changed set to 0.

Fix that by setting color_mgmt_changed to true in
drm_atomic_helper_check_modeset instead of setting it when the
gamma_lut/degamma_lut/ctm are set in the new crtc_state.

Changes since v2:
  - Instead of setting color_mgmt_changed in commit_duplicated_set
    just set it in check_modeset and clean up other place where it was
    setted, suggested by Maarten Lankhorst.

Change-Id: I7afe6d14fa9ab5bc9b5a314730011e95c754a0e7
Signed-off-by: Alexandru Gheorghe <alexandru-cosmin.gheorghe@arm.com>