PM / devfreq: stopping the governor before device_unregister()
device_release() is freeing the resources before calling the device
specific release callback which is, in the case of devfreq, stopping
the governor.

It is a problem as some governors are using the device resources. e.g.
simpleondemand which is using the devfreq deferrable monitoring work. If it
is not stopped before the resources are freed, it might lead to a use after
free.

Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>
Reviewed-by: John Einar Reitan <john.reitan@arm.com>
[cw00.choi: Fix merge conflict]
Reviewed-by: Chanwoo Choi <cw00.choi@samsung.com>
Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>

(cherry picked from commit 2f061fd0c2d852e32e03a903fccd810663c5c31e from
mainline)

Jira: MPDBSP-2056
Change-Id: I9fe73fedfc9a6b8bcdad147e4bd8ce90ca26ab97
Signed-off-by: Steven Wang <steven.wang@arm.com>