staging: android: ion: flush cache after allocation from kernel page pool
The memory allocated from the kernel page pool is likely to be in the CPU cache.
It might have been used previously and because of the GFP_ZERO flag,
alloc_pages() will write it. In order to map this memory as non-cached, we need
to flush the CPU cache first, to avoid any memory corruption when the
corresponding cache line gets evicted.

A proper fix would be to flush the cache using the Linux DMA API. The problem is
that the buffer is not always attached to a device and then, in the case of
arm64, any dma_sync*() call would be a no-op.

This patch is then a quick dirty fix, calling directly arm64 cache flush
function in ION to ensure that the cache line is actually cleaned before the
memory is used as non-cached. The cache gets invalidated as well as this will
diverge from the actual memory.

Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>
(cherry picked from commit 8dbe047cb2e637de0a51799422924cb87a535327)

The ion_page_pool struct does not use the 'cached' member anymore. Thus an extra
'cached' parameter is added to related function prototypes to distinguish between
cached and uncached buffers.

Jira: MPDBSP-2056
Change-Id: I8ae7b586ce624c1a6a2d2d45df2440860fbd3001
Signed-off-by: Steven Wang <steven.wang@arm.com>