staging: android: ion: flush cache after allocation from kernel page pool
The memory allocated from the kernel page pool is likely to be in the CPU cache.
It might have been used previously and because of the GFP_ZERO flag,
alloc_pages() will write it. In order to map this memory as non-cached, we need
to flush the CPU cache first, to avoid any memory corruption when the
corresponding cache line gets evicted.

A proper fix would be to flush the cache using the Linux DMA API. The problem is
that the buffer is not always attached to a device and then, in the case of
arm64, any dma_sync*() call would be a no-op.

This patch is then a quick dirty fix, calling directly arm64 cache flush
function in ION to ensure that the cache line is actually cleaned before the
memory is used as non-cached. The cache gets invalidated as well as this will
diverge from the actual memory.

Jira: MPDBSP-1680
(ported from commit 8dbe047cb2e637de0a51799422924cb87a535327)
Change-Id: Iaf803b74ed6b7edde4bbc9813c7207332f8c9a1f
Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>