staging: android: ion fix pool->cached member set
This fix is in drivers/staging/android/ion, and is introduced from
commit 0699d92a092a246aa0b20a5b403643e25a60e0f0 of
mpg-bsp.cambridge.arm.com/bsp/linux.git repo.

The pool struct is initialized with kmalloc, the value of any of its member is
then random and it's unlikely that "cached" will be set to false by default. All
members needs to be properly initialized.

Change-Id: Ie6526359f1d1380e5df89bcb38a2cf3584a0f3d3