staging: android: ion: flush cache after allocation from kernel page pool
This fix is in drivers/staging/android/ion, and is introduced from
commit b371d110b254364fc5bd8692f3cee219193b869f of
mpg-bsp.cambridge.arm.com/bsp/linux.git repo.

The memory allocated from the kernel page pool is likely to be in the CPU cache.
It might have been used previously and because of the GFP_ZERO flag,
alloc_pages() will write it. In order to map this memory as non-cached, we need
to flush the CPU cache first, to avoid any memory corruption when the
corresponding cache line gets evicted.

A proper fix would be to flush the cache using the Linux DMA API. The problem is
that the buffer is not always attached to a device and then, in the case of
arm64, any dma_sync*() call would be a no-op.

This patch is then a quick dirty fix, calling directly arm64 cache flush
function in ION to ensure that the cache line is actually cleaned before the
memory is used as non-cached. The cache gets invalidated as well as this will
diverge from the actual memory.

Change-Id: I3e36c22a4eb1f55241be9950856ca0083d182cc4