drm: Add command mode support to our virtual-encoder
We have introduced a new field (bool command_mode)  in 'struct drm_display_info'.
This denotes that the display controller is in single frame refresh mode.

One needs to parse for the property "command_mode" in dts node to determine if
the display controller needs to be in single frame refresh mode.

Change-Id: I27249a6682fab4bb5663e9045067f9e6aab8d668
Signed-off-by: Ayan Kumar halder <ayan.halder@arm.com>