DISPLAYSW-3172: drm/komeda: Force dirty changed flags when enable a plane
komeda only update komeda state when crtc/plane is active, so once
the user do the update during the inactive period, we may lose the
changed flag in the active period, and lose the drm state updating
that in the inactive period.
To avoid such problem, force dirty all the changed flags when switch
a plane from disable to enable.

Change-Id: I3e79f3825ae6af0d08660b739f146bb49aee3b41
Signed-off-by: james qian wang (Arm Technology China) <james.qian.wang@arm.com>