Revert "drm/komeda: Consolidate struct komeda_drv allocations"
This reverts commit e1c46bcf26d4bbf34c7a76b33f293d552d91c69c.

The drm_device absolutely needs to be freed and recreated on every
bind-unbind cycle, so this is actually going in the wrong direction
memory-management wise, at least at this moment. We should be able to
move the mm once we drop bind/unbind and use only probe/remove (which
coincides with going all bridge-and-device-links.

Change-Id: I8c1d51f7812150a954f2ec5f3c5e98333cc41abc